test: move serializer-shape tests to test_serializers.py
The response-envelope, sealed-secret-field, and manifest-serializer
tests assert serializer output shape, so group them with the other
serializer tests. test_schema_hook.py keeps the routing and
generated-schema-component tests.
