+++ apps/api/export/tests/test_schema_hook.py
@@ -5,18 +5,10 @@ subclass carrying that schema."""
 import pytest
 from django.urls import resolve
 from drf_spectacular.generators import SchemaGenerator
-from rest_framework import serializers
 
-from apps.api.export.serializers import (
-    ManifestEntrySerializer,
-    ManifestSerializer,
-    build_resource_response_serializer,
-    build_resource_serializer,
-    component_name,
-    resource_responses,
-)
+from apps.api.export.serializers import component_name, resource_responses
 from apps.api.export.views import ResourceView, resource_view
-from apps.teams.export.manifest import MANIFEST_ENTRIES, build_manifest, entry_model, get_manifest_entry
+from apps.teams.export.manifest import MANIFEST_ENTRIES, entry_model, get_manifest_entry
 
 
 def test_each_resource_resolves_to_a_resourceview_carrying_its_name():
@@ -41,12 +33,6 @@ def test_unknown_resource_resolves_to_catchall_view():
     assert match.kwargs == {"resource": "not_a_real_resource"}
 
 
-def test_response_envelope_has_cursor_has_more_and_results():
-    fields = build_resource_response_serializer(get_manifest_entry("users"))().fields
-    assert set(fields) == {"cursor", "has_more", "results"}
-    assert isinstance(fields["results"], serializers.ListSerializer)
-
-
 @pytest.mark.parametrize(
     ("resource", "expect_409"),
     [
@@ -60,23 +46,6 @@ def test_secret_resources_document_the_no_public_key_conflict(resource, expect_4
     assert 200 in statuses
     assert 404 not in statuses
     assert (409 in statuses) is expect_409
-
-
-def test_secret_fields_documented_as_sealed_strings():
-    # llm_providers' `config` is sealed to a base64 string at runtime, not its raw model type.
-    model = entry_model(get_manifest_entry("llm_providers").model)
-    fields = build_resource_serializer(model)().fields
-    assert isinstance(fields["config"], serializers.CharField)
-
-
-@pytest.mark.django_db()
-def test_manifest_serializer_matches_build_manifest_payload():
-    """The documented manifest response must stay in step with what the view actually returns."""
-    manifest = build_manifest()
-    assert set(ManifestSerializer().fields) == set(manifest)
-    declared = set(ManifestEntrySerializer().fields)
-    actual_keys = set().union(*(entry.keys() for entry in manifest["entries"]))
-    assert declared == actual_keys
 
 
 @pytest.fixture(scope="module")
